# filap-api/api_spec.yml
basePath: ''
consumes:
- application/json
...
host: localhost:5000
info:
  contact:
    name: Filap API Support
  description: A minimal live Q&A API for hosts and audiences
  termsOfService: ''
  title: Filap API
  version: 1.0.0
...
paths:
...
  /api/queues:
    post:
      consumes:
      - application/json
      parameters:
      - in: body
        name: body
        schema:
          properties:
            default_sort_order:
              default: votes
              description: Default sorting order for messages
              enum:
              - votes
              - newest
              type: string
            name:
              description: Optional queue name
              type: string
          type: object
      responses:
        '201':
          description: Queue created successfully
          schema:
            properties:
              created_at:
                format: date-time
                type: string
              default_sort_order:
                enum:
                - votes
                - newest
                type: string
              expires_at:
                format: date-time
                type: string
              host_secret:
                description: Host authentication secret
                format: uuid
                type: string
              id:
                description: Unique queue identifier
                format: uuid
                type: string
              name:
                description: Queue name
                type: string
            type: object
        '400':
          description: Invalid request data
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new queue
      tags:
      - Queues
  /api/queues/{queue_id}:
    get:
      parameters:
      - description: Queue identifier
        format: uuid
        in: path
        name: queue_id
        required: true
        type: string
      responses:
        '200':
          description: Queue metadata
          schema:
            properties:
              default_sort_order:
                description: Default sorting order
                enum:
                - votes
                - newest
                type: string
              expires_at:
                description: Queue expiration time
                format: date-time
                type: string
              id:
                description: Queue identifier
                format: uuid
                type: string
              name:
                description: Queue name
                type: string
            type: object
        '404':
          description: Queue not found or expired
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get queue metadata and settings
      tags:
      - Queues
    patch:
      parameters:
      - description: Queue identifier
        format: uuid
        in: path
        name: queue_id
        required: true
        type: string
      - description: Host authentication secret
        format: uuid
        in: header
        name: X-Queue-Secret
        required: true
        type: string
      - in: body
        name: body
        schema:
          properties:
            default_sort_order:
              description: Default sorting order for messages
              enum:
              - votes
              - newest
              type: string
            name:
              description: Updated queue name
              type: string
          type: object
      responses:
        '200':
          description: Queue updated successfully
          schema:
            properties:
              default_sort_order:
                enum:
                - votes
                - newest
                type: string
              expires_at:
                format: date-time
                type: string
              id:
                format: uuid
                type: string
              name:
                type: string
            type: object
        '400':
          description: Invalid request data
          schema:
            properties:
              error:
                type: string
            type: object
        '401':
          description: Missing host secret
          schema:
            properties:
              error:
                type: string
            type: object
        '403':
          description: Invalid host credentials
          schema:
            properties:
              error:
                type: string
            type: object
        '404':
          description: Queue not found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - HostSecret: []
      summary: Update queue settings (host only)
      tags:
      - Queues
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  HostSecret:
    in: header
    name: X-Queue-Secret
    type: apiKey
  UserToken:
    in: header
    name: X-User-Token
    type: apiKey
swagger: '2.0'